+++ bump-version.sh
@@ -14,7 +14,7 @@ fi
 # CHANGELOG.md
 VER=$1
 TODAY=`date +%F`
-${SED} -i.bak -e "/SNAPSHOT/c\
+${SED} -i -e "/SNAPSHOT/c\
 ## ${VER} / ${TODAY}" CHANGELOG.md
 
 # clj