+++ src/chat/client.cljs
@@ -1,5 +1,6 @@
 (ns chat.client
   (:require [cljs.core.async :as a :refer [<! >! go go-loop]]
+            [cljs-http.client :as http]
             [clojure.string :as str]
             [haslett.client :as ws]
             [haslett.format :as wsfmt]))
@@ -7,12 +8,25 @@
 (defn- query [query]
   (.querySelector js/document query))
 
+(defn author []
+  (.-value (query "#author")))
+
+(defn- admin? []
+  (= (author) "hkimura"))
+
+(defn- abbrev [s]
+  (str (first s) "*****"))
+
 (defn- append-html [element html]
+  ;; https://qiita.com/isseium/items/12b215b6eab26acd2afe
+  (.play js/sound)
   (.insertAdjacentHTML element "afterbegin" html))
 
 (defn- message-html [{:keys [author message]}]
-  (str "<li><span class='author'>"
-       (if (str/blank? author) "Anonymous" author) "</span>"
+  (str "<li><span class='date'>"
+       (js/Date.)
+        "<br><span class='author'>"
+       (if (str/blank? author) "Anonymous" (abbrev author)) "</span>"
        "<span class='message'>" message "</span></li>"))
 
 (defn- send-message [stream]
@@ -52,6 +66,36 @@
   (js/console.log "websocket-connect")
   (ws/connect (websocket-url path) {:format wsfmt/transit}))
 
+(defn- insert-random-user []
+  (go (let [response (<! (http/get "/api/user-random"))
+            user (:user (:body response))]
+         ;; this is it!
+        (set! (.-value (query "#message")) (str "@" user " ")))))
+
+;; from biff,
+;; (map message (sort-by :msg/sent-at #(compare %2 %1) messages))
+(defn- format-message [{:keys [author message timestamp]}]
+  (str "<p>&nbsp" timestamp "<br>&nbsp"
+       "<b>" (abbrev author) ":</b> " message "</p>"))
+
+(defn- replace-content [element messages]
+  ;; this again. do not forget.
+  (set! (.-textContent element) "")
+  (doseq [msg (sort-by :timestamp #(compare %1 %2) messages)]
+    (.insertAdjacentHTML element "afterbegin" (format-message msg))))
+
+(defn- remove? [msg]
+  (let [m (:message msg)]
+    (if (str/starts-with? m (str "@" (author) " "))
+      false
+      (str/starts-with? m "@"))))
+
+(defn- load-messages [n]
+  (go (let [response (<! (http/get (str "/api/load/" n)))
+            messages (:body response)]
+        (replace-content
+         (query "#message-log") (remove remove? messages)))))
+
 (defn- on-load [_]
   (js/console.log "on-load")
   (go (let [stream  (<! (websocket-connect "/chat"))
@@ -62,11 +106,15 @@
         (.focus message)
         (.addEventListener message "keyup"
                            (fn [e]
-                             (when (and
-                                    (= (.-code e) "Enter")
-                                    (.-shiftKey e))
-                               (js/console.log (str e))
-                               (send-message stream)))))))
+                             (cond
+                               (and (= (.-code e) "Enter") (.-shiftKey e))
+                               (send-message stream)
+                               (and (= (.-code e) "KeyU") (.-ctrlKey e))
+                               (if (admin?)
+                                 (insert-random-user)
+                                 (js/alert "admin only.")))))
+        (.addEventListener (query "#load") "click"
+                           (fn [_] (load-messages 10))))))
 
 (defn init []
   (js/console.log "init")