+++ src/chat/server.clj
@@ -3,25 +3,32 @@
   (:require [buddy.hashers :as hashers]
             [clojure.core.async :as a]
             [clojure.java.io :as io]
+            [java-time.api :as jt]
             [clojure.string :as str]
             [hato.client :as hc]
+            [muuntaja.middleware :as mw]
             [reitit.ring :as rr]
             [ring.adapter.jetty :as adapter]
             [ring.util.anti-forgery :refer [anti-forgery-field]]
             [ring.middleware.defaults :as def]
             [ring.util.response :as resp]
-            [ring.websocket.async :as wsa]
+            ;; [ring.websocket.async :as wsa]
+            [chat.async :as wsa]
             [ring.websocket.transit :as wst]
             [ring.websocket.keepalive :as wska]
-            [taoensso.telemere :as t]))
+            [taoensso.telemere :as t]
+            ;;
+            [chat.xtdb :as xt]))
 
-(t/set-min-level! (if (System/getenv "MX3_DEV") :debug :info))
-;; (t/log! {:level :debug :id "telemere"} "debug level")
-;; (t/log! :info  ["telemere" "info" "level"])
+(def debug? (System/getenv "MX3_DEV"))
+(t/set-min-level! (if debug? :debug :info))
 
-(def ^:private version "v0.10.88")
+(def ^:private version "v0.15.163")
 
-(def ^:pricate url "https://l22.melt.kyutech.ac.jp/api/user/")
+(def ^:private l22
+  (if debug?
+    "http://localhost:3090/"
+    "https://l22.melt.kyutech.ac.jp/"))
 
 (defn make-chat-handler []
   (let [writer  (a/chan)
@@ -36,7 +43,7 @@
     (-> (resp/response
          (str
           "<!DOCTYPE html><title>MX3</title>
-           <h1>Micro X version3</h1>
+           <h1>Micro X for Classes</h1>
            <body style='font-family:sans-serif;'>
            <form method='post'>"
           (anti-forgery-field)
@@ -52,11 +59,12 @@
         (resp/charset "UTF-8"))))
 
 (defn login! [{{:keys [login password]} :params}]
-  (if (System/getenv "MX3_DEV")
+  (if debug?
     (-> (resp/redirect "/index")
         (assoc-in [:session :identity] login))
     (try
-      (let [resp (hc/get (str url login) {:timeout 3000 :as :json})]
+      (let [resp (hc/get (str l22 "api/user/"login)
+                         {:timeout 3000 :as :json})]
         (if (and (some? resp)
                  (hashers/check password (get-in resp [:body :password])))
           (-> (resp/redirect "/index")
@@ -78,11 +86,60 @@
           (resp/content-type "text/html")
           (resp/charset "UTF-8")))))
 
+;; must be rewritten with java-time. agry.
+(defn- utime [t]
+  (cond
+    debug? "1"
+    (< (+ (* 8 60) 50) t  (+ (* 10 60) 20)) "1"
+    (< (+ (* 10 60) 30) t (+ (* 12 60))) "2"
+    :else "0"))
+
+(defn- uhour []
+  (let [[wd _ _ hhmmss] (-> (str (java.util.Date.))
+                            (str/split #"\s"))
+        [hh mm] (str/split hhmmss #":")
+        t (+ (* 60 (Long/parseLong hh)) (Long/parseLong mm))]
+    (str/lower-case (str
+                     (if debug? "wed" wd)
+                     (utime t)))))
+
+(defn user-random [_]
+  (-> (hc/get (str l22 "api/user/" (uhour) "/randomly")
+              {:as :json :timeout 1000})
+      :body))
+
+(defn- load-data [{{:keys [n]} :path-params}]
+  (xt/q '{:find [author message timestamp]
+          :keys [author message timestamp]
+          :in [t0]
+          :where [[e :author author]
+                  [e :message message]
+                  [e :timestamp timestamp]
+                  [(<= t0 timestamp)]]}
+        (jt/minus (jt/local-date-time) (jt/minutes (Long/parseLong n)))))
+
+;; (defn- load-data [{{:keys [n]} :path-params :as request}]
+;;   (def *r* request)
+;;   (xt/q '{:find [(pull eid [*])]
+;;           ;; :keys [author message timestamp]
+;;           :in [t0]
+;;           :where [[eid :timestamp timestamp]
+;;                   [(<= t0 timestamp)]]}
+;;         (jt/minus (jt/local-date-time) (jt/minutes (Long/parseLong n)))))
+
+
 (defn make-app-handler []
   (rr/ring-handler
    (rr/router [["/chat" {:middleware [[wst/wrap-websocket-transit]
                                       [wska/wrap-websocket-keepalive]]}
                 ["" (make-chat-handler)]]
+               ["/api" {:middleware [[def/wrap-defaults def/api-defaults]
+                                     mw/wrap-format
+                                     mw/wrap-params]}
+                ["/load/:n" (fn [req]
+                              (resp/response (load-data req)))]
+                ["/user-random" {:get (fn [_]
+                                        (resp/response (user-random nil)))}]]
                ["" {:middleware [[def/wrap-defaults def/site-defaults]]}
                 ["/" {:get login :post login!}]
                 ["/logout" (fn [_]
@@ -93,6 +150,8 @@
     (rr/create-resource-handler {:path "/"})
     (rr/create-default-handler))
    {:middleware []}))
+
+;;
 
 (defn run-server [options]
   (adapter/run-jetty (make-app-handler) options))
@@ -106,12 +165,14 @@
   ([{:keys [port]}]
    (when-not (some? @server)
      (reset! server (run-server {:port port :join? false}))
-     (println "server started in port " port "."))))
+     (xt/start! "config.edn")
+     (println "server started in port" port))))
 
 (defn stop []
   (when (some? @server)
     (.stop @server)
     (reset! server nil)
+    (xt/stop!)
     (println "server stopped.")))
 
 (defn restart []
@@ -122,4 +183,8 @@
   (start))
 
 (comment
-  (restart))
+  (xt/q '{:find [who what when]
+          :where [[e :author who]
+                  [e :message what]
+                  [e :timestamp when]]})
+  :rcf)